refactor(cli): split portal.rs into portal/ module (closes #40)
Break the 3070-line crates/harn-cli/src/commands/portal.rs into a module
tree so each file has a single responsibility and no file exceeds ~760
lines. The HTTP API surface and JSON shapes are byte-identical; all 22
existing portal tests are preserved in portal/tests.rs.

Layout:
- dto.rs        — ~30 Portal* DTO structs + MaterializedLaunchTarget
- query.rs      — RunQuery / CompareQuery / ListRunsQuery / ErrorResponse
- state.rs      — PortalState
- assets.rs     — PORTAL_DIST + index/asset route fns + content-type map
- errors.rs     — bad_request/not_found/internal_error helpers
- util.rs       — compact_json, format_duration, preview_text, etc.
- highlight.rs  — /api/highlight/keywords builder
- llm.rs        — /api/llm/options + provider model discovery
- transcript.rs — JSONL transcript step parsing + story sections
- run_analysis.rs — scan_runs, filter_and_sort, build_run_detail +
                    all stage/span/activity/transition/checkpoint/
                    artifact/policy/replay/stage-debug sub-builders
- launch.rs     — launch job materialization + target scanning
- router.rs     — build_router
- handlers/     — one file per route family (runs, compare, meta, launch)
- tests.rs      — all existing portal tests, unchanged

Verified via cargo check -p harn-cli, cargo test -p harn-cli (130/130),
portal binary smoke test, npm run portal:{lint,test,build}, and make all
(472 conformance + 330 docs snippets).
