+++ crates/harn-cli/src/commands/portal/errors.rs
@@ -1,0 +1,31 @@
+use axum::http::StatusCode;
+use axum::Json;
+
+use super::query::ErrorResponse;
+
+pub(super) fn internal_error(message: impl ToString) -> (StatusCode, Json<ErrorResponse>) {
+    (
+        StatusCode::INTERNAL_SERVER_ERROR,
+        Json(ErrorResponse {
+            error: message.to_string(),
+        }),
+    )
+}
+
+pub(super) fn bad_request_error(message: impl ToString) -> (StatusCode, Json<ErrorResponse>) {
+    (
+        StatusCode::BAD_REQUEST,
+        Json(ErrorResponse {
+            error: message.to_string(),
+        }),
+    )
+}
+
+pub(super) fn not_found_error(message: impl ToString) -> (StatusCode, Json<ErrorResponse>) {
+    (
+        StatusCode::NOT_FOUND,
+        Json(ErrorResponse {
+            error: message.to_string(),
+        }),
+    )
+}
